refactor the Scope class, extract import and export components logic to new files, as well as getting the scope remotes in order to break a circle dependency of scope => remotes => remote => connect => fs => loadScope => scope
