+++ src/api/scope/lib/modify-ci-props.js
@@ -2,10 +2,12 @@
 import { loadScope, Scope } from '../../../scope';
 import { BitId } from '../../../bit-id';
 import type ConsumerComponent from '../../../consumer/component';
+import ScopeComponentsImporter from '../../../scope/component-ops/scope-components-importer';
 
 export default (async function modifyCIProps(path: string, id: string, ciProps: Object): Promise<any> {
   const scope: Scope = await loadScope(path);
   const bitId: BitId = await scope.getParsedId(id);
-  const component: ConsumerComponent = await scope.loadComponent(bitId);
+  const scopeComponentsImporter = ScopeComponentsImporter.getInstance(scope);
+  const component: ConsumerComponent = await scopeComponentsImporter.loadComponent(bitId);
   return scope.sources.modifyCIProps({ source: component, ciProps });
 });
