+++ src/search/search-adapter.js
@@ -4,6 +4,7 @@ import indexer from './indexer';
 import { search } from './searcher';
 import { loadConsumer } from '../consumer';
 import { loadScope } from '../scope';
+import { getScopeRemotes } from '../scope/scope-remotes';
 
 async function searchLocally(queryStr: string, reindex: boolean = false): Promise<any> {
   let scopePath;
@@ -27,7 +28,7 @@ async function searchLocally(queryStr: string, reindex: boolean = false): Promis
 
 async function searchRemotely(queryStr: string, scope: string, reindex: boolean = false): Promise<any> {
   return loadConsumer().then((consumer) => {
-    return consumer.scope.remotes().then(remotes =>
+    return getScopeRemotes(consumer.scope).then(remotes =>
       remotes.resolve(scope, consumer.scope.name).then((remote) => {
         return remote.search(queryStr, reindex);
       })
